\name{NegBinBetaBinreg-package}
\alias{NegBinBetaBinreg-package}
\docType{package}
\title{NegBinBetaBinreg}
\description{Function to estimate a Negative Binomial regression models with mean and shape (or variance) regression structures, and  Beta Binomial regression with mean and dispersion regression structures.}
\details{
\tabular{ll}{
Package: \tab NegBinBetaBinreg \cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-10-8\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Edilberto Cepeda-Cuervo \email{ecepedac@unal.edu.co},
 Maria Victoria Cifuentes-Amado \email{mvcifuentesa@unal.edu.co},
 Margarita Marin \email{mmarinj@unal.edu.co}  
}
\keyword{package}




